Revert "Remove --local argument to pcs setup"
This reverts commit 22f7fb528f258b5fcc0e954462dbfe2484992dd1.

Removing the --local argument from "pcs cluster setup" causes pacemaker
to attempt to connect to pcsd on port 2224 for each of the listed nodes.

pcsd is a service which is used to bootstrap the cluster, syncing the
pacemaker and corosync configuration files out to the cluster nodes and
starting corosync and pacemaker (and perhaps other tasks; see man pcsd).

We don't use pcsd since we use cfengine to sync out configurations, so
nothing is listening on that port, the connection attempt fails, the
"pcs cluster setup" fails, and ldev2pcs fails.

So we need the "--local" argument.

Fixes: https://lc.llnl.gov/jira/browse/TOSS-4146